requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.87.0
fileid==0.2.1
filetype==1.2.0
Flask==2.2.2
flatbuffers==22.12.6
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.4
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.1
lxml==4.9.1
MarkupSafe==2.1.1
mpmath==1.2.1
multidict==6.0.3
Naked==0.1.32
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==22.0
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.10
pycryptodome==3.16.0
pydantic==1.10.2
PyMatting==1.1.8
pyreadline3==3.4.1
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0
rembg==2.0.29
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.21.0
sympy==1.11.1
tifffile==2022.10.10
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
uuid==1.30
uvicorn==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
yarl==1.8.2
